WheelFile: memoize distinfo directory prefix
This has to be done because corrupted files in lazy mode do not
necessarily follow the convention implemented in `_distinfo_path()`.

With this change, the prefix can be found outside `_distinfo_path()`, or
generated inside it once, if required.

Then, other subroutines can reuse the prefix without needlessly
regenerating it every time, or creating a duplicate distinfo directory
in a wheel where the prefix is different.