WheelFile: refactor creation from unnamed objects
If WheelFile is given a directory or a buffer as its first argument, it
has to infer the wheel attributes (name, version, tags) from the rest of
the arguments.

Previously this was done by introducing separate branch early in the
__init__ call. This was unnecessary. Most of that branch is now moved to
all of the separate _pick_* methods, which makes for a more readable
code.