Preserve ZipInfo in writestr_* and from_wheelfile
The following `WheelFile` methods will now preserve `ZipInfo` attributes
if a `ZipInfo` object is given to them:

- `writestr_data`
- `writestr_distinfo`

Previously they have not been doing so.

The `WheelFile.from_wheelfile` constructor will now also preserve
distinfo attributes for the files inside data and distinfo archive
directories.

Thanks to Maxime Boissonneault (GH @mboisson) for the code he sent in
his pull request for this, which served as a guidepost for this
implementation's tests and some of the details.